[feat] Refresh Token 발급, Refresh Token을 통한 Access Token 재발급 기능 구현

+++ src/main/java/com/encore/logeat/common/jwt/refresh/UserRefreshTokenRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRefreshTokenRepository extends JpaRepository<UserRefreshToken, Long> {
 
-	Optional<UserRefreshToken> findByUserIdAndReissueCountLessThan(Long id, int count);
+	Optional<UserRefreshToken> findByUserIdAndReissueCountLessThan(Long id, long count);
 
 }
